+++ changelog.md
@@ -11,8 +11,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 - [Added compiler plugin validations for Postgresql as a datasource](https://github.com/ballerina-platform/ballerina-library/issues/5829)
+- [Added compiler plugin validations for new advanced SQL database annotations](https://github.com/ballerina-platform/ballerina-library/issues/6068)
 
 ### Changed
+- [Changed the behavior of foreign key presence validation to account for Relation annotation](https://github.com/ballerina-platform/ballerina-library/issues/6068)
 
 ## [1.1.0] - 2023-06-30